This now runs under Linux GCC. Still haven't tested on Windows

+++ src/Utils.hpp
@@ -4,6 +4,7 @@
 #include <ostream>
 #include <random>
 #include <cassert>
+#include <cfloat>
 
 static const int MAX_RAY_DEPTH = 35;
 const float PI = 3.14159265358979323846;
@@ -50,7 +51,7 @@ struct vec3 {
 	}
 
 	float length() const {
-		return std::sqrtf(lengthSquare());
+		return std::sqrt(lengthSquare());
 	}
 
 	float lengthSquare() const {
